Also Allowing diagonal movements
This is the modified version of the original one. It also considers diagonally adjacent blocks as neighbors. So 2 classes of neighbors, one with distance 1 unit and another one with 1.414 unit. So changes are made to select non-barrier neighbors as well. 
This version runs quite faster than the previous one, allowing movement along the x and y axes only.